Fixed an error about json format and added a new route to the urls file
So the problem was that json didn't accept dictionary keys to be date time object so using the strftime method,I converted the datetime object to string and passed it as key.
As for the urls file i added a route that will contain data brought from the database in a json format using the jsonify method of the flask package.